TDS-1435:  Rename SupportToolProperties class (and associated settings)
